+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features Added
 - Added support for service version 2021-06-08.
+- Added a new paginated method for listing page ranges, `list_page_ranges()`. This replaces `get_page_ranges()` which has been deprecated.
 - Added support for `copy_source_tags` to `start_copy_from_url()` which allows callers to specify whether blob tags
   should be replaced or copied during a synchronous copy operation.
 - Added support for `create_if_not_exists()` for `BlobContainerClient`